+++ CHANGELOG.rst
@@ -8,6 +8,7 @@ master
 - Minor changes to logging
 - install command now runs script in verbose mode for better debugging once I leave the ship
 - Append current log to expeditioner email
+- Added more commands to run on unmount
 
 0.2.4
 -----